re: #233 Refactor formatdef_eslint_local
The old formatdef for eslint could not correctly locate the binary for
eslint installed via `sudo npm -g install eslint`. It would also not
work if a `eslintrc.*` file did not exist.

This simple patch makes the formatter fall back on using the $PATH to
locate `eslint` if it is not a locally installed node_module, allows the
formatter to run with the --no-eslintrc option if no config file can be
found, and simplifies some of the logic where the formatter attempts to
find the config files.

Patch tested with `vim` 8.0 compiled for Fedora 28 Workstation